backend/requirements.txt
fastapi>=0.115.0,<1.0.0
fastapi-utils>=0.8.0,<1.0.0
uvicorn>=0.30.0,<1.0.0
sqlalchemy>=2.0.23,<3.0.0
psycopg2-binary>=2.9.10,<3.0.0
pydantic>=2.10.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
openai>=1.3.5,<2.0.0
reportlab>=4.0.0,<5.0.0
passlib>=1.7.4,<2.0.0
# passlib 1.7.x bcrypt handler is not compatible with bcrypt>=4.
bcrypt>=3.2.2,<4.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
python-multipart>=0.0.9,<1.0.0
httpx>=0.24.0,<1.0.0
alembic>=1.11.1,<2.0.0
